Revert "Remove prompt to delete failed training data file"
This reverts commit ba1694f26e29ae0cc580d1d698fc6a2b64d8969d.

+++ src/tumblrbot/flow/fine_tune.py
@@ -8,6 +8,7 @@ from currency_converter import CurrencyConverter
 from rich import print as rich_print
 from rich.console import Console
 from rich.progress import open as progress_open
+from rich.prompt import Confirm
 from tiktoken import encoding_for_model, get_encoding
 
 from tumblrbot.utils.common import FlowClass, PreviewLive, localize_number
@@ -95,12 +96,12 @@ class FineTuner(FlowClass):
         self.config.job_id = ""
 
         if job.status != "succeeded":
-            rich_print("[gray62]Fine-tuning failed. Deleting the uploaded examples file...")
-            self.openai.files.delete(job.training_file)
-            rich_print()
+            if Confirm.ask("[gray62]Delete uploaded examples file?", default=False):
+                self.openai.files.delete(job.training_file)
+                rich_print()
 
-            if job.status == "failed" and job.error is not None:
-                raise RuntimeError(job.error.message)
+            message = "Fine-tuning failed!" if job.error is None else job.error.message
+            raise RuntimeError(message or "Fine-tuning cancelled!")
 
         if job.fine_tuned_model is not None:
             self.config.fine_tuned_model = job.fine_tuned_model
